fix: merge visualize e2e tests into root_session_test for pkg-file-count
